requirements.txt
absl-py==1.4.0
aiml==0.9.2
asgiref==3.7.2
asttokens==2.0.8
astunparse==1.6.3
backcall==0.2.0
cachetools==5.3.0
certifi==2022.6.15
charset-normalizer==2.1.0
colorama==0.4.5
cycler==0.11.0
decorator==5.1.1
distlib==0.3.7
executing==1.0.0
filelock==3.12.2
flatbuffers==23.3.3
fonttools==4.33.3
gast==0.4.0
google-auth==2.17.3
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.53.0
h5py==3.8.0
idna==3.3
imgviz==1.7.2
ipython==8.5.0
jax==0.4.8
jedi==0.18.1
joblib==1.2.0
keras==2.12.0
kiwisolver==1.4.2
labelme==5.2.0.post4
libclang==16.0.0
Markdown==3.4.3
MarkupSafe==2.1.2
matplotlib==3.5.2
matplotlib-inline==0.1.6
ml-dtypes==0.1.0
natsort==8.3.1
numpy==1.22.4
oauthlib==3.2.2
opencv-python==4.8.0.76
opt-einsum==3.3.0
packaging==21.3
pandas==1.5.2
parso==0.8.3
pickleshare==0.7.5
Pillow==9.1.1
platformdirs==3.10.0
prompt-toolkit==3.0.31
protobuf==4.22.3
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pygame==2.1.2
Pygments==2.13.0
pyparsing==3.0.9
PyQt5==5.15.6
PyQt5-Qt5==5.15.2
PyQt5-sip==12.10.1
python-aiml==0.9.3
python-dateutil==2.8.2
pytz==2022.7
PyYAML==6.0
pyzmq==24.0.1
QtPy==2.3.1
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
scikit-learn==1.2.2
scipy==1.8.1
six==1.16.0
sqlparse==0.4.4
stack-data==0.5.0
tensorboard==2.12.2
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.8.1
tensorflow==2.12.0
tensorflow-estimator==2.12.0
tensorflow-intel==2.12.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.2.0
threadpoolctl==3.1.0
torch==1.12.1
torchvision==0.13.1
traitlets==5.3.0
typing_extensions==4.3.0
tzdata==2023.3
urllib3==1.26.11
vboxapi==1.0
virtualenv==20.24.2
wcwidth==0.2.5
Werkzeug==2.2.3
wrapt==1.14.1
